requirements.txt
# Dependencias principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0

# Utilidades
pydantic==2.5.0
pydantic-settings==2.1.0
httpx==0.25.2

# Desarrollo
python-decouple==3.8